Phase 3: Fix printf and small warning categories (C4474/C4477/C4778/C4244/C4838/C4018) (#3)
* yml変更

* fix: add explicit double cast for position format specifiers (C4477)


* fix: correct printf format specifiers in FileRetentionPeriod.cpp (C4474/C4477/C4778)


* fix: add LPCTSTR casts for remaining CString sprintf calls (C4477)


* fix: correct unterminated format strings in MainFrm.cpp (C4474/C4778)


* fix: correct format specifier from %lf to %d for int type (C4477)


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - batch 1


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - batch 2


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - batch 3


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - batch 4


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - PatRecDeskewDlg


* fix: add LPCTSTR casts for CString in sprintf calls (C4477) - Recipe files


* fix: add explicit casts for narrowing conversions (C4244/C4838)


* fix: cast to DWORD for signed/unsigned comparison (C4018)


---------

+++ SRC/ConfigFile/IniFile.cpp
@@ -1332,7 +1332,7 @@ void LoadNanoSpecIni(int iIniFile)
 
 // 2009.09.29 bagus SE --{--
 	if (iIniFile == SYSTEM_SR_INI_FILE) {
-		NanoSysData.SrConfig.wTrShutterMoveWaitTime = lg_TrShutterMoveWaitTime * 1000;
+		NanoSysData.SrConfig.wTrShutterMoveWaitTime = static_cast<WORD>(lg_TrShutterMoveWaitTime * 1000);
 	}
 // 2009.09.29 bagus SE --}--
 
@@ -1396,7 +1396,7 @@ void LoadNanoSpecIni(int iIniFile)
 			NanoSysData.SrConfig.dUvGain					= 0.000;
 			NanoSysData.SrConfig.bTransmittance				= FALSE;
 			NanoSysData.SrConfig.iTrUseLensNo				= 0;	// Turret No:1
-			NanoSysData.SrConfig.wTrShutterMoveWaitTime		= 0.000;
+			NanoSysData.SrConfig.wTrShutterMoveWaitTime		= static_cast<WORD>(0);
 			NanoSysData.SrConfig.bGantryType				= FALSE;
 			NanoSysData.SrConfig.wGantryNumOfMeasPoint		= 1;
 			NanoSysData.SrConfig.bDistance					= FALSE;
@@ -1636,7 +1636,7 @@ void SaveNanoSpecIni(int iIniFile)
 			if (NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn < 1)
 				continue;
 
-			for(nIniStressSection = 0;nIniStressSection < NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn - 1;nIniStressSection++){
+			for(nIniStressSection = 0;static_cast<DWORD>(nIniStressSection) < NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn - 1;nIniStressSection++){
 // 2009.09.17 bagus SE --}--
 				sprintf(szBuf,"L%dS%d_ScanStart_X",nIniStressLine+1,nIniStressSection+1);
 				wsprintf(buff,"%ld", NanoSysData.StressConfig.Line[nIniStressLine].SectPos[nIniStressSection].lScanStartPosX);
